Fixes all of the warnings in the .log supplied by Richard (@hobbes1069).. hopefully.

+++ src/webp.imageio/webpinput.cpp
@@ -82,7 +82,7 @@ WebpInput::open (const std::string &name, ImageSpec &spec)
 
     std::vector<uint8_t> encoded_image;
     encoded_image.resize(m_image_size, 0);
-    int numRead = fread(&encoded_image[0], sizeof(uint8_t), encoded_image.size(), m_file);
+    size_t numRead = fread(&encoded_image[0], sizeof(uint8_t), encoded_image.size(), m_file);
     if (numRead != encoded_image.size()) {
     	error ("Did not read correct amount of images. (expected %d, read %d)",
     			encoded_image.size(), numRead);
